requirements.txt
# FastAPI Backend
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9
sqlalchemy>=2.0.28

# LlamaIndex & Vector DB
llama-index>=0.10.20
llama-index-llms-ollama>=0.1.2
llama-index-embeddings-huggingface>=0.1.3
chromadb>=0.4.24
qdrant-client>=1.8.0

# Data Processing
pandas>=2.2.1
openpyxl>=3.1.2

# Enterprise Multi-Agent & UI
streamlit>=1.32.0
langgraph>=0.0.30
langchain>=0.1.13
langchain-google-genai>=1.0.1
python-dotenv>=1.0.1
langchain-community>=0.0.29
duckduckgo-search>=5.2.1
beautifulsoup4>=4.12.3
requests>=2.31.0
fake-useragent>=1.5.1
tenacity>=8.2.3
